MIDI Loader rewrite
