+++ src/spessasynth_lib/midi_parser/midi_sequence.js
@@ -124,9 +124,9 @@ export class MIDISequenceData
     format = 0;
     
     /**
-     * The RMID (Resource Interchangeable MIDI) info data, if the file is RMID formatted.
+     * The RMID (Resource-Interchangeable MIDI) info data, if the file is RMID formatted.
      * Otherwise, this field is undefined.
-     * Chunk type (e.g. "INAM"): Chunk data as binary array.
+     * Chunk type (e.g. "INAM"): Chunk data as a binary array.
      * @type {Object<string, IndexedByteArray>}
      */
     RMIDInfo = {};
